+++ iamirror/MirrorAppOptions.cpp
@@ -83,7 +83,7 @@ bool AppOptions::initaliseConfig() {
 	std::string temp = SAUtils::GetPOSIXEnv(key);
 	std::string homePath = temp;
 	//printf("%s", homePath.c_str());
-	int i = homePath.length();
+	size_t i = homePath.length();
 	if (homePath.empty() == true || homePath.length() == 0) {
 		homePath = SAUtils::GetPOSIXEnv("ProgramData");
 		//C:\ProgramData\IDK Software\ImageArchive1.0
