+++ util/LogDocument.cpp
@@ -77,7 +77,7 @@ bool LogDocument::write(FormatType formatType) {
 
 bool iequals(const std::string& a, const std::string& b)
 {
-	unsigned int sz = a.size();
+	size_t sz = a.size();
 	if (b.size() != sz)
 		return false;
 	for (unsigned int i = 0; i < sz; ++i)
